UI rewrite (Blade design system) + DB-backed user management

+++ admin/admin-ajax.php
@@ -18,24 +18,44 @@ switch( $action ) {
 
     case 'add':
         yourls_verify_nonce( 'add_url', $_REQUEST['nonce'], false, 'omg error' );
-        $return = yourls_add_new_link( $_REQUEST['url'], $_REQUEST['keyword'], '', $_REQUEST['rowid'] );
+        if ( function_exists( 'yourls_current_user_can' ) && !yourls_current_user_can( 'create_link' ) ) {
+            yourls_status_header( 403 );
+            echo json_encode( [ 'status' => 'fail', 'message' => yourls__( 'Forbidden' ) ] );
+            die();
+        }
+        $return = yourls_add_new_link( $_REQUEST['url'], $_REQUEST['keyword'], '', $_REQUEST['rowid'] ?? 1, $_REQUEST['notes'] ?? '' );
         echo json_encode($return);
         break;
 
     case 'edit_display':
         yourls_verify_nonce( 'edit-link_'.$_REQUEST['id'], $_REQUEST['nonce'], false, 'omg error' );
+        if ( function_exists( 'yourls_current_user_can' ) && !yourls_current_user_can( 'edit_link', [ 'keyword' => $_REQUEST['keyword'] ?? '' ] ) ) {
+            yourls_status_header( 403 );
+            echo json_encode( [ 'status' => 'fail', 'message' => yourls__( 'Forbidden' ) ] );
+            die();
+        }
         $row = yourls_table_edit_row ( $_REQUEST['keyword'], $_REQUEST['id'] );
         echo json_encode( array('html' => $row) );
         break;
 
     case 'edit_save':
         yourls_verify_nonce( 'edit-save_'.$_REQUEST['id'], $_REQUEST['nonce'], false, 'omg error' );
-        $return = yourls_edit_link( $_REQUEST['url'], $_REQUEST['keyword'], $_REQUEST['newkeyword'], $_REQUEST['title'] );
+        if ( function_exists( 'yourls_current_user_can' ) && !yourls_current_user_can( 'edit_link', [ 'keyword' => $_REQUEST['keyword'] ?? '' ] ) ) {
+            yourls_status_header( 403 );
+            echo json_encode( [ 'status' => 'fail', 'message' => yourls__( 'Forbidden' ) ] );
+            die();
+        }
+        $return = yourls_edit_link( $_REQUEST['url'], $_REQUEST['keyword'], $_REQUEST['newkeyword'], $_REQUEST['title'], $_REQUEST['notes'] ?? null );
         echo json_encode($return);
         break;
 
     case 'delete':
         yourls_verify_nonce( 'delete-link_'.$_REQUEST['id'], $_REQUEST['nonce'], false, 'omg error' );
+        if ( function_exists( 'yourls_current_user_can' ) && !yourls_current_user_can( 'delete_link', [ 'keyword' => $_REQUEST['keyword'] ?? '' ] ) ) {
+            yourls_status_header( 403 );
+            echo json_encode( [ 'status' => 'fail', 'message' => yourls__( 'Forbidden' ) ] );
+            die();
+        }
         $query = yourls_delete_link_by_keyword( $_REQUEST['keyword'] );
         echo json_encode(array('success'=>$query));
         break;
